+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed: Save multiple Show related post values
 * Fixed: Count bug on Override Archive shortcode
 * Fixed: Date/time display on Override Archive shortcode
+* Fixed: Show file Disable Download checkbox saving
+* Added: Use automatic Embeds on external Show file URLs
 * Changed: Removed player state saving iframe
 
 = 2.5.7 =